requirements.txt
discord.py
psycopg2-binary
swgoh_comlink
titlecase
unidecode
asyncpg
async-comlink